Разбивает скрипты на модули
